+++ NEWS.md
@@ -16,6 +16,8 @@
 
 * Added `vec()`: efficiently turn matrices or data frames / lists into a single atomic vector. I am aware of multiple implementations in other packages, which are mostly inefficient. With atomic objects, `vec()` simply removes the attributes without copying the object, and with lists it directly calls `C_pivot_longer`. 
 
+* Added `fmatch()`: a striking fast `match()` function. 
+
 ### Improvements
 
 * `set_collapse()` now supports options 'mask' and 'remove', giving *collapse* a flexible namespace in the broadest sense that can be changed at any point within the active session: